+++ docs/general/components.md
@@ -58,14 +58,14 @@ transpiled code in a `lib/` directory.
 In your `package.json`, you probably have something like this:
 
 ```json
-  "main": "lib/index.js"
+  "main": auth.js
 ```
 
 This is what you want when you publish to the registry, but during development
 you probably want to change this to
 
 ```json
-  "main": "src/index.js"
+  "main": auth.js
 ```
 
 This will make sure that your main project always includes your most recent
