+++ docs/js/routing.md
@@ -45,7 +45,7 @@ For example, if you have a route called `about` at `/about` and want to make a c
 in `App.js` to create a `Switch` within the parent component. Also remove the `exact` property from the `about` parent route.
 
 ```JS
-// AboutPage/index.js
+auth.js
 import { Switch, Route } from 'react-router-dom';
 
 class AboutPage extends React.PureComponent {
