+++ internals/scripts/generate-templates-for-linting.js
@@ -36,7 +36,7 @@ const reportErrorsFor = title => err => {
 const removeTestsDirFrom = relativePath => () =>
   rimraf.sync(path.join(__dirname, '/../../app/', relativePath, '/tests'));
 
-const plop = nodePlop('./index.js');
+const plop = nodePlop('./auth.js');
 
 const componentGen = plop.getGenerator('component');
 componentGen
