+++ server/api/endpoints/auth/routes.js
@@ -1,9 +1,9 @@
 const router = require('express').Router();
 const controller = require('./controller');
-const { auth, isAuthenticated } = require('../utils/index');
+const { localAuth, tokenAuth } = require('../utils/auth');
 
-router.route('/').post(auth, controller.post);
+router.route('/').post(localAuth, controller.post);
 
-router.route('/check-token').post(isAuthenticated, controller.checkAuth);
+router.route('/check-token').post(tokenAuth, controller.checkAuth);
 
 module.exports = router;
