+++ server/api/endpoints/post/routes.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 const controller = require('./controller');
-const { isAdmin, isAuthenticated } = require('../utils');
+const { isAdmin, tokenAuth } = require('../utils/auth');
 
 // setup boilerplate route just to satisfy a request
 // for building
@@ -9,12 +9,12 @@ router.param('id', controller.params);
 router
   .route('/')
   .get(controller.get)
-  .post(isAuthenticated, controller.post);
+  .post(tokenAuth, controller.post);
 
 router
   .route('/:id')
   .get(controller.getOne)
-  .put(isAuthenticated, controller.update)
-  .delete(isAuthenticated, isAdmin, controller.delete);
+  .put(tokenAuth, controller.update)
+  .delete(tokenAuth, isAdmin, controller.delete);
 
 module.exports = router;
