+++ server/api/endpoints/user/routes.js
@@ -1,7 +1,7 @@
 const router = require('express').Router();
 
 const controller = require('./controller');
-const { isAdmin, isAuthenticated } = require('../utils');
+const { isAdmin, tokenAuth } = require('../utils/auth');
 
 // setup boilerplate route just to satisfy a request
 // for building
@@ -10,7 +10,7 @@ router.param('id', controller.params);
 
 router
   .route('/')
-  .get(isAuthenticated, isAdmin, controller.get)
+  .get(tokenAuth, isAdmin, controller.get)
   .post(controller.post);
 
 router
