+++ server/api/endpoints/utils/auth.js
@@ -1,11 +1,11 @@
 const passport = require('passport');
 
 module.exports = {
-  auth: passport.authenticate('local', {
+  localAuth: passport.authenticate('local', {
     usernameField: 'email',
     session: false,
   }),
-  isAuthenticated: passport.authenticate('jwt', { session: false }),
+  tokenAuth: passport.authenticate('jwt', { session: false }),
   isAdmin: (req, res, next) => {
     if (req.user && req.user.admin) {
       return next();
